refactored app.
